refactor (front): renomeando a NavLayout -> PageLayout
